library/python/django_abc_data/requirements.txt
django-appconf==1.0.2
ids[abc]==1.3.66
sync-tools==0.3.1
django-closuretree==1.2.0
tvmauth==3.2.3
